pyproject.toml
[project]
name = "retrieval-weight-experiment"
version = "0.1.0"
description = "Gradient-free retrieval weight optimization via Thompson Sampling with LLM self-assessment"
authors = [{name = "Alfonso DiRocco", email = "alfonso@kusp.dev"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "anthropic>=0.40",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "requests>=2.31",
    "langfuse>=2.0",
    "matplotlib>=3.10.8",
    "scipy>=1.11",
]

[project.urls]
Repository = "https://github.com/kusp-dev/retrieval-weight-experiment"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.15",
    "pre-commit>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"future/*" = ["E402"]
"scripts/*" = ["E501"]

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true
check_untyped_defs = false
exclude = ["future/", "thompson-sampling/", "tests/"]
